requirements.txt
astroid==2.12.12
autopep8==2.0.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
dill==0.3.6
et-xmlfile==1.1.0
Flask==2.2.2
Flask-CKEditor==0.4.6
idna==3.4
importlib-metadata==5.0.0
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.8.0
MarkupSafe==2.1.1
mccabe==0.7.0
numpy==1.23.4
openai==0.24.0
openpyxl==3.0.10
pandas==1.5.1
pandas-stubs==1.5.1.221024
platformdirs==2.5.2
pycodestyle==2.9.1
pylint==2.15.5
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.5
requests==2.28.1
six==1.16.0
tomli==2.0.1
tomlkit==0.11.6
tqdm==4.64.1
types-pytz==2022.5.0.0
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
WTForms==3.0.1
zipp==3.10.0
